Struct waveguide::engine::Engine

pub struct Engine {

    cur_time: f64,
    ex: ScalarField,
    ey: ScalarField,
    ez: ScalarField,
    hx: ScalarField,
    hy: ScalarField,
    hz: ScalarField,
    hx_wtr: Writer<BufWriter<File>>,
    hy_wtr: Writer<BufWriter<File>>,
    hz_wtr: Writer<BufWriter<File>>,
    ex_wtr: Writer<BufWriter<File>>,
    ey_wtr: Writer<BufWriter<File>>,
    ez_wtr: Writer<BufWriter<File>>,
    snapshot_steps: usize,

}

Engine struct

contains all code and data necessary to evolve field states to a given time

Fields
cur_time: f64ex: ScalarFieldey: ScalarFieldez: ScalarFieldhx: ScalarFieldhy: ScalarFieldhz: ScalarFieldhx_wtr: Writer<BufWriter<File>>hy_wtr: Writer<BufWriter<File>>hz_wtr: Writer<BufWriter<File>>ex_wtr: Writer<BufWriter<File>>ey_wtr: Writer<BufWriter<File>>ez_wtr: Writer<BufWriter<File>>snapshot_steps: usize
Implementations

impl Engine

pub fn new(config: &Config, geometry: &Geometry) -> Result<Engine>

Engine constructor

Arguments

config solver::Config struct

geometry geometry::Geometry struct

Returns

Result<Engine>

Errors

any ScalarField constructors error -fs::create_dir() is unable to create directory “./out”

creation of any file descriptors fail

pub fn update(&mut self, config: &Config, geometry: &Geometry) -> Result<()>

updates all fields to time specified in config.end_time

Arguments

config solver::Config struct

geometry geometry::Geometry struct

Returns

Result<()>

Errors

any snapshot_fields() call errors

flush_fields() call errors

fn snapshot_fields(&mut self) -> Result<()>

snapshots all field values

Arguments

&mut self mutable reference to Engine struct

Returns

Result<()>

Errors

write_buf_vec() calls fail for any field

fn flush_fields(&mut self) -> Result<()>

flushes remaining data in buffers to disk if present

Arguments

&mut self mutable reference to Engine struct

Returns

Result<()>

Errors

wtr.flush() calls fail for any field

fn write_buf_vec( buffered_writer: &mut Writer<BufWriter<File>>, data: &[f64] ) -> Result<()>

writes data found in &f64 into a buffered writer

Arguments

buffered_writer &mut Writer<BufWriter<File>> mutable reference to a buffered writer

data &f64 array slice to be written to disk

Returns

Result<()>

Errors

buffered_writer() call fails

fn update_h(&mut self, geometry: &Geometry, hax: &f64, hay: &f64, haz: &f64)

updates all magnetic field components in dielectric region using standard 3D Yee scheme

Arguments

&mut self &mut Engine mutable reference to Engine struct

geometry &Geometry reference to a Geometry struct

hax &f64 reference to Hx update coefficient

hay &f64 reference to Hy update coefficient

haz &f64 reference to Hz update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls inside of any field update equations

fn update_hx(&mut self, geometry: &Geometry, hay: &f64, haz: &f64)

updates all Hx components in dielectric region using standard 3D Yee scheme

Arguments

&mut self &mut Engine mutable reference to Engine struct

geometry &Geometry reference to a Geometry struct

hay &f64 reference to Hy update coefficient

haz &f64 reference to Hz update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn update_hy(&mut self, geometry: &Geometry, hax: &f64, haz: &f64)

updates all Hy components in dielectric region using standard 3D Yee scheme

Arguments

&mut self &mut Engine mutable reference to Engine struct

geometry &Geometry reference to a Geometry struct

hax &f64 reference to Hx update coefficient

haz &f64 reference to Hz update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn update_hz(&mut self, geometry: &Geometry, hax: &f64, hay: &f64)

updates all Hz components in dielectric region using standard 3D Yee scheme

Arguments

&mut self &mut Engine mutable reference to Engine struct

geometry &Geometry reference to a Geometry struct

hax &f64 reference to Hx update coefficient

hay &f64 reference to Hy update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn update_e(&mut self, geometry: &Geometry, ea: &f64, eb: &f64)

updates all electric field components in dielectric region using standard 3D Yee scheme

Arguments

&mut self &mut Engine mutable reference to Engine struct

geometry &Geometry reference to a Geometry struct

ea &f64 reference electric field ‘alpha’ update coefficient

eb &f64 reference electric field ‘beta’ update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls inside of any field update equations

fn update_ex(&mut self, geometry: &Geometry, ea: &f64, eb: &f64)

updates all Ex components in dielectric region using standard 3D Yee scheme

Arguments

&mut self &mut Engine mutable reference to Engine struct

geometry &Geometry reference to a Geometry struct

ea &f64 reference electric field ‘alpha’ update coefficient

eb &f64 reference electric field ‘beta’ update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn update_ey(&mut self, geometry: &Geometry, ea: &f64, eb: &f64)

updates all Ey components in dielectric region using standard 3D Yee scheme

Arguments

&mut self &mut Engine mutable reference to Engine struct

geometry &Geometry reference to a Geometry struct

ea &f64 reference electric field ‘alpha’ update coefficient

eb &f64 reference electric field ‘beta’ update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn update_ez(&mut self, geometry: &Geometry, ea: &f64, eb: &f64)

updates all Ez components in dielectric region using standard 3D Yee scheme

Arguments

&mut self &mut Engine mutable reference to Engine struct

geometry &Geometry reference to a Geometry struct

ea &f64 reference electric field ‘alpha’ update coefficient

eb &f64 reference electric field ‘beta’ update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn update_tfsf_source_hx( &mut self, config: &Config, geometry: &Geometry, hay: &f64 )

injects an Ey source field into Hx using the TF/SF method with for waveguide source at z-high plane

Arguments

&mut self &mut Engine mutable reference to Engine struct

config &Config reference to a Config struct

geometry &Geometry reference to a Geometry struct

hay &f64 reference to Hy update coefficient

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn update_tfsf_source_ey( &mut self, config: &Config, geometry: &Geometry, dt: &f64 )

injects an Hx source field into Ey using the TF/SF method for waveguide source at z-high plane

Arguments

&mut self &mut Engine mutable reference to Engine struct

config &Config reference to a Config struct

geometry &Geometry reference to a Geometry struct

dt &f64 reference to time step

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn tapered_sin(&self, config: &Config) -> f64

computes the value of a tapered sin wave using parameters defined in Config struct

Arguments

self &Engine reference to Engine struct

config &Config reference to a Config struct

Returns

Errors

fn update_mur_abc_ey( &mut self, geometry: &Geometry, abc1: &f64, abc2: &f64, abc3: &f64, abc4: &f64, ey_n: &ScalarField, ey_n_1: &ScalarField )

updates Ey field points on z-low plane using Mur ABC scheme

Arguments

&mut self &mut Engine mutable reference to Engine structt

geometry &Geometry reference to a Geometry struct

abc1 &f64 reference to 1st Mur ABC update coefficient

abc2 &f64 reference to 2nd Mur ABC update coefficient

abc3 &f64 reference to 3rd Mur ABC update coefficient

abc4 &f64 reference to 4th Mur ABC update coefficient

ey_n &ScalarField reference to cloned Ey field at t=n

ey_n_1 &ScalarField reference to cloned Ey field at t=n-1

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls

fn update_mur_abc_hx( &mut self, geometry: &Geometry, abc1: &f64, abc2: &f64, abc3: &f64, abc4: &f64, hx_n: &ScalarField, hx_n_1: &ScalarField )

updates Hx field points on z-low plane using Mur ABC scheme

Arguments

&mut self &mut Engine mutable reference to Engine structt

geometry &Geometry reference to a Geometry struct

abc1 &f64 reference to 1st Mur ABC update coefficient

abc2 &f64 reference to 2nd Mur ABC update coefficient

abc3 &f64 reference to 3rd Mur ABC update coefficient

abc4 &f64 reference to 4th Mur ABC update coefficient

hx_n &ScalarField reference to cloned Hx field at t=n

hx_n_1 &ScalarField reference to cloned Hx field at t=n-1

Returns

Errors

Will not error, will panic at out of bounds accesses in ScalarField.idx() or ScalarField.idxm() calls