Struct waveguide::engine::ScalarField

struct ScalarField {
    field: Vec<f64>,
    row_offset: usize,
    column_offset: usize,
}

ScalarField struct

Provides a no cost abstraction of a Vec<f64> to function as a 3D scalar field using Fortran-style ordering as this intuitively simpler to implement than C-style in my opinion

Fields
field: Vec<f64>row_offset: usizecolumn_offset: usize
Implementations

impl ScalarField

pub fn new( initial_value: f64, size: usize, row_offset: usize, column_offset: usize ) -> Result<ScalarField>

ScalarField constructor

Arguments

initial_value [implied unit] initial value for all points in ScalarField

size [] total number of discrete locations in field

row_offset [] row offset into 1D Vec

column_offset [] column offset into 1D Vec

Returns

Result<ScalarField>

Errors

pub fn idx(&self, i: usize, j: usize, k: usize) -> f64

returns an aliased value at position [i,j,k] in ScalarField

used to get an immutable value of a ScalarField at position [i,j,k]

Arguments

-i [] i-index into ScalarField using Fortran-style ordering -j [] j-index into ScalarField using Fortran-style ordering -k [] k-index into ScalarField using Fortran-style ordering

Returns

f64 aliased value at position [i,j,k] in ScalarField

Errors

panics if [i,j,k] is out of bounds. this was not error handled properly to reduce the number of checks as this function is used very heavily

pub fn idxm(&mut self, i: usize, j: usize, k: usize) -> &mut f64

returns a mutable reference to value at position [i,j,k] in ScalarField

used to get a mutable reference to a value of a ScalarField at position [i,j,k]

Arguments

-i [] i-index into ScalarField using Fortran-style ordering -j [] j-index into ScalarField using Fortran-style ordering -k [] k-index into ScalarField using Fortran-style ordering

Returns

&mut f64 mutable reference to value at position [i,j,k] in ScalarField

Errors

panics if [i,j,k] is out of bounds. this was not error handled properly to reduce the number of checks as this function is used very heavily