Struct waveguide::solver::Config

pub struct Config {

    pub max_frequency: f64,
    pub voxels_per_min_wavelength: usize,
    pub voxels_per_min_feature: usize,
    pub x_len: f64,
    pub y_len: f64,
    pub z_len: f64,
    pub ep_r: f64,
    pub mu_r: f64,
    pub sigma: f64,
    pub snapshot_steps: usize,
    pub buffered_snapshots: usize,
    pub end_time: f64,
    pub frequency: f64,
    pub delay_time: f64,
    pub ramp_time: f64,

}

Config struct

contains all data imported from ./config.toml which is used to bootstrap other simulation structures

if I had more time, I would rewrite other structures to not use this struct past the bootstrapping stage

Fields
max_frequency: f64voxels_per_min_wavelength: usizevoxels_per_min_feature: usizex_len: f64y_len: f64z_len: f64ep_r: f64mu_r: f64sigma: f64snapshot_steps: usizebuffered_snapshots: usizeend_time: f64frequency: f64delay_time: f64ramp_time: f64
Implementations

impl Config

pub fn new(path_str: &str) -> Result<Config>

Config constructor

Arguments

path_str &str path to desired configuration file

Returns

Result<Config>

Errors

path_str is not able to be converted to PathBuf

path_str is not a valid path on the filesystem

path_str does not correspond to a file

file at path_str cannot be read into a String

file string from path_str cannot be deserialized into toml